LibGfx: Move ICCProfile.{h,cpp} to ICC/Profile.{h,cpp}
